requirements.txt
python-telegram-bot==20.3
openai==0.11.0
requests==2.31.0
python-dotenv==1.0.0